more beaker logic
